api/requirements.txt
eth_account==0.13.7
fastapi==0.120.1
itsdangerous
joblib==1.5.1
matplotlib==3.8.4
numpy==1.23.2
pandas==1.5.3
psycopg2-binary==2.9.10
pydantic==2.12.3
PyJWT==2.10.1
Requests==2.32.5
schemas==0.7.1
scikit_learn==1.7.2
SQLAlchemy==2.0.38
starlette==0.49.1
uvicorn==0.23.2
python-dotenv
apscheduler
bcrypt
mlflow
paho-mqtt
aiohttp
asyncpg
alembic